+++ kifi-backend/modules/search/app/com/keepit/search/graph/bookmark/URIGraphSearcher.scala
@@ -29,9 +29,6 @@ import com.keepit.search.graph.LuceneBackedEdgeSet
 import com.keepit.search.graph.URIList
 import com.keepit.search.graph.Util
 import com.keepit.search.graph.bookmark.URIGraphFields._
-import com.keepit.search.graph.LuceneBackedBookmarkTimeRangeFilter
-import com.keepit.search.graph.TimeRangeFilter
-import com.keepit.search.graph.BookmarkInfoAccessor
 
 object URIGraphSearcher {
   def apply(uriGraphIndexer: URIGraphIndexer): URIGraphSearcher = {
@@ -39,7 +36,7 @@ object URIGraphSearcher {
     new URIGraphSearcher(indexSearcher, storeSearcher)
   }
 
-  def apply(userId: Id[User], uriGraphIndexer: URIGraphIndexer, shoeboxClient: ShoeboxServiceClient, monitoredAwait: MonitoredAwait): URIGraphSearcherWithUser = {
+  def apply(userId: Id[User], uriGraphIndexer: URIGraphIndexer, shoeboxClient: ShoeboxServiceClient, monitoredAwait: MonitoredAwait): URIGraphSearcher = {
     val (indexSearcher, storeSearcher) = uriGraphIndexer.getSearchers
     new URIGraphSearcherWithUser(indexSearcher, storeSearcher, userId, shoeboxClient, monitoredAwait)
   }
@@ -143,7 +140,7 @@ class URIGraphSearcherWithUser(searcher: Searcher, storeSearcher: Searcher, myUs
   def getBookmarkRecord(uriId: Id[NormalizedURI]): Option[BookmarkRecord] = {
     import com.keepit.search.graph.bookmark.BookmarkRecordSerializer._
 
-    val bookmarkId = myUriEdgeSet.getBookmarkId(uriId.id)
+    val bookmarkId = myUriEdgeSet.accessor.getBookmarkId(uriId.id)
     storeSearcher.getDecodedDocValue[BookmarkRecord](BookmarkStoreFields.recField, bookmarkId)
   }
 }
@@ -185,13 +182,13 @@ object UserToUserEdgeSet{
   }
 }
 
-abstract class UserToUriEdgeSet(override val sourceId: Id[User]) extends BookmarkInfoAccessor[User, NormalizedURI] with TimeRangeFilter[User, NormalizedURI]
+abstract class UserToUriEdgeSet(override val sourceId: Id[User]) extends EdgeSet[User, NormalizedURI]
 
 object UserToUriEdgeSet {
   def apply(sourceId: Id[User], uriList: URIList, isPublicEdgeSet: Boolean): UserToUriEdgeSet = {
     val set = LongArraySet.fromSorted(uriList.ids)
 
-    new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] with LuceneBackedBookmarkTimeRangeFilter[User, NormalizedURI]{
+    new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] {
       override protected val longArraySet = set
       override protected def createdAtByIndex(idx:Int): Long = {
         val datetime = uriList.createdAt(idx)
@@ -211,7 +208,7 @@ object UserToUriEdgeSet {
       val set = LongArraySet.from(concat(publicIds, privateIds))
       val pubListSize = publicIds.length
 
-      new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] with LuceneBackedBookmarkTimeRangeFilter[User, NormalizedURI] {
+      new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] {
         override protected val longArraySet = set
         override protected def createdAtByIndex(idx:Int): Long = {
           val datetime = if (idx < pubListSize) publicList.createdAt(idx) else privateList.createdAt(idx - pubListSize)
@@ -230,7 +227,7 @@ object UserToUriEdgeSet {
     val set = LongArraySet.from(myInfo.uriIdArray, myInfo.mapper.reserveMapper)
 
     val pubListSize = publicList.size
-    new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] with LuceneBackedBookmarkTimeRangeFilter[User, NormalizedURI]{
+    new UserToUriEdgeSet(sourceId) with LongSetEdgeSetWithAttributes[User, NormalizedURI] {
       override protected val longArraySet = set
       override protected def createdAtByIndex(idx:Int): Long = {
         val datetime = if (idx < pubListSize) publicList.createdAt(idx) else privateList.createdAt(idx - pubListSize)
