+++ kifi-backend/modules/search/test/com/keepit/search/graph/EdgeSetTest.scala
@@ -30,16 +30,16 @@ class EdgeSetTest extends Specification {
       var timeArray = Array[Long](100, 200, 300, 110, 120, 130, 101, 102, 103)
       var edgeSet = new TestEdgeSet(Id[Vertex](0), LongArraySet.fromSorted(idArray), timeArray)
 
-//      edgeSet.filterByTimeRange(100, 130).destIdLongSet === Set[Long](1, 4, 5, 6, 7, 8, 9)
-//      edgeSet.filterByTimeRange(101, 119).destIdLongSet === Set[Long](4, 7, 8, 9)
-//
-//      // unsorted ids (reversed arrays)
-//      idArray = idArray.reverse
-//      timeArray = timeArray.reverse
-//      edgeSet = new TestEdgeSet(Id[Vertex](0), LongArraySet.from(idArray), timeArray)
-//
-//      edgeSet.filterByTimeRange(100, 130).destIdLongSet === Set[Long](1, 4, 5, 6, 7, 8, 9)
-//      edgeSet.filterByTimeRange(101, 119).destIdLongSet === Set[Long](4, 7, 8, 9)
+      edgeSet.filterByTimeRange(100, 130).destIdLongSet === Set[Long](1, 4, 5, 6, 7, 8, 9)
+      edgeSet.filterByTimeRange(101, 119).destIdLongSet === Set[Long](4, 7, 8, 9)
+
+      // unsorted ids (reversed arrays)
+      idArray = idArray.reverse
+      timeArray = timeArray.reverse
+      edgeSet = new TestEdgeSet(Id[Vertex](0), LongArraySet.from(idArray), timeArray)
+
+      edgeSet.filterByTimeRange(100, 130).destIdLongSet === Set[Long](1, 4, 5, 6, 7, 8, 9)
+      edgeSet.filterByTimeRange(101, 119).destIdLongSet === Set[Long](4, 7, 8, 9)
     }
   }
 }
